Fixed issue #1994
